Added GuzzleClient
