+++ composer.json
@@ -19,6 +19,9 @@
     "require-dev": {
         "phpunit/phpunit": "~4.0"
     },
+    "suggest": {
+      "guzzlehttp/guzzle": "Allows the usage of the provided GuzzleClient instead of CURL"
+    },
     "autoload": {
         "psr-0": {
             "Redmine": "lib/"
